Move library files into pkg/wol
